% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyomics_colours.R
\name{polyomics_colours}
\alias{polyomics_colours}
\title{Polyomics colours}
\usage{
polyomics_colours()
}
\value{
Vector of all approved Polyomics colours and their hex values
}
\description{
List of approved Glasgow Polyomics colours
}
\examples{
# Get the polyomics colours
cols <- polyomics_colours()

# Display Polyomics colours as a colour chart with colur name and hex value
n_seq <- seq_along(cols)
image(1, n_seq, t(as.matrix(n_seq)), col = cols,
      xlab = "", ylab = "", xaxt = "n", yaxt = "n", bty = "n",
      main = "Corporate colors for Glasgow Polyomics")
text(0.8, n_seq, names(cols), col = "white")
text(1.2, n_seq, cols, col = "white")

}
