% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyomics_colours_hex.R
\name{polyomics_cols}
\alias{polyomics_cols}
\title{Polyomics colours hexadecimal list}
\usage{
polyomics_cols(...)
}
\arguments{
\item{...}{Passed in colours}
}
\value{
List of colours and their hexadecimal values
}
\description{
Polyomics colours hexadecimal list
}
