% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyomics_palettes.R
\name{polyomics_pal}
\alias{polyomics_pal}
\title{Polyomics Palettes}
\usage{
polyomics_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{NAmed colour palette to use}

\item{reverse}{Reverse the order of the colour list}

\item{...}{Params set to NULL}
}
\value{
Returns a list of colours from a set colour palette list
}
\description{
Returns a list of colours from a set colour palette list
}
\examples{
# Get the list of Polyomics colours
polyomics_colors <- polyomics_colours()
# Get the list of Polyomics colour palettes
polyomics_palettes <- polyomics_palettes_list()

# Set the number of colours to retrieve
levels <- 3

# Set the palette to use
cols <- polyomics_pal("hot")(levels)

# Display the colour palettes
n_seq <- seq_along(cols)
image(n_seq, 1, as.matrix(n_seq), col = cols,
      xlab = "", ylab = "", xaxt = "n", yaxt = "n", bty = "n",
      main = "Cool palette of polyomics Inc")
text(n_seq, 1, cols, col = "white", srt = 90)

}
