% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyomics_palette_list.R
\name{polyomics_palettes_list}
\alias{polyomics_palettes_list}
\title{Polyomics Colour Palettes}
\usage{
polyomics_palettes_list(...)
}
\arguments{
\item{...}{Params set to NULL}
}
\value{
List of approved Polyomics colour palettes
}
\description{
Polyomics Colour Palettes
}
\examples{
polyomics_palettes <- polyomics_palettes_list()
polyomics_palettes
}
