% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyomics_scale_colour.R
\name{polyomics_scale_colour}
\alias{polyomics_scale_colour}
\title{Polyomics Scale Colour}
\usage{
polyomics_scale_colour(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in polyomics_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean to reverse the order of the colour palette}

\item{...}{Params set to NULL}
}
\value{

}
\description{
Color scale constructor for polyomics colors
}
\examples{
library(ggplot2)
plot <- ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) +
polyomics_scale_color(palette = "mixed")

}
