% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyomics_scale_fill.R
\name{polyomics_scale_fill}
\alias{polyomics_scale_fill}
\title{Fill scale constructor for Polyomics colors}
\usage{
polyomics_scale_fill(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in polyomics_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for Polyomics colors
}
\examples{
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
 polyomics_scale_fill(palette = "grey", guide = "none")

}
